% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset}
\alias{get_dataset}
\alias{has_dataset}
\alias{clear_dataset}
\title{Get, check or clear the default dataset}
\usage{
get_dataset()

has_dataset()

clear_dataset()
}
\value{
\code{get_dataset()} returns the default data frame (error if none is set);
\code{has_dataset()} returns a logical; \code{clear_dataset()} returns \code{TRUE}
invisibly.
}
\description{
Companions to \code{\link[=use_dataset]{use_dataset()}} for inspecting and resetting the session's
default dataset.
}
\details{
\code{has_dataset()} is the safe way to check before calling \code{get_dataset()}, which
errors when nothing is set. \code{clear_dataset()} removes the default so that
subsequent calls require an explicit \code{data} again -- useful at the end of a
script or between analyses of different datasets.
}
\examples{
has_dataset()
#> [1] FALSE

use_dataset(nps_drivers)
has_dataset()
#> [1] TRUE

nrow(get_dataset())
#> [1] 600

clear_dataset()
}
\seealso{
\code{\link[=use_dataset]{use_dataset()}}.

Other config:
\code{\link[=edit_ezrmodel_profile]{edit_ezrmodel_profile()}},
\code{\link[=ezrmodel_options]{ezrmodel_options()}},
\code{\link[=ezrmodel_profile_paths]{ezrmodel_profile_paths()}},
\code{\link[=load_ezrmodel_profile]{load_ezrmodel_profile()}},
\code{\link[=reset_ezrmodel_options]{reset_ezrmodel_options()}},
\code{\link[=save_ezrmodel_profile]{save_ezrmodel_profile()}},
\code{\link[=use_dataset]{use_dataset()}},
\code{\link[=use_ezrmodel_profile]{use_ezrmodel_profile()}}
}
\concept{config}
